+++ src/Pages/Proposal/Components/index.js
@@ -2,3 +2,5 @@ export { default as ProposalDeposits } from './ProposalDeposits';
 export { default as ProposalInformation } from './ProposalInformation';
 export { default as ProposalTimingProgressBar } from './ProposalTimingProgressBar';
 export { default as ProposalTimingTable } from './ProposalTimingTable';
+export { default as ProposalVotingGraph } from './ProposalVotingGraph';
+export { default as ProposalVotingTable } from './ProposalVotingTable';