+++ src/Pages/Proposal/Proposal.js
@@ -7,6 +7,8 @@ import {
   ProposalInformation,
   ProposalTimingProgressBar,
   ProposalTimingTable,
+  ProposalVotingGraph,
+  ProposalVotingTable,
 } from './Components';
 
 const Proposal = () => {
@@ -36,6 +38,12 @@ const Proposal = () => {
       </Section>
       <Section>
         <ProposalDeposits />
+      </Section>
+      <Section>
+        <ProposalVotingGraph />
+      </Section>
+      <Section>
+        <ProposalVotingTable />
       </Section>
     </Wrapper>
   );