+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Add Gov link to global nav #135
 - Add proposal detail page and proposal info #136
 - Add proposal timing detail and progress bar #137
+- Add proposal voting detail and progress bar #138
 - Add proposol deposit card to detail page #139
 
 ### Improvements